Move AudioController to audio package
